fix(#76): Remove unnecessary unboxing operations
The change eliminates unnecessary unboxing in the sliding window algorithm for the minimum window substring problem. By directly working with Integer objects in the frequency maps, the code becomes cleaner and more efficient. This change addresses potential performance issues and improves readability.

+++ LeetCode/76-minimum-window-substring/Solution.java
@@ -47,7 +47,7 @@ public class Solution {
 
         // Remove character from the left from the window
         windowCounts.put(c, windowCounts.get(c) - 1);
-        if (targetFreq.containsKey(c) && windowCounts.get(c).intValue() < targetFreq.get(c).intValue()) {
+        if (targetFreq.containsKey(c) && windowCounts.get(c) < targetFreq.get(c)) {
           formed--;
         }
         left++;
